requirements.txt
backports.functools-lru-cache==1.5
biopython==1.73
cycler==0.10.0
decorator==4.3.2
funcsigs==1.0.2
GridDataFormats==0.4.0
gsd==1.6.0
h5py==2.9.0
joblib==0.13.2
kiwisolver==1.0.1
matplotlib==2.2.3
MDAnalysis==0.19.2
mmtf-python==1.1.2
mock==2.0.0
msgpack==0.6.1
networkx==2.2
numpy==1.16.2
pbr==5.1.2
pyparsing==2.3.1
python-dateutil==2.8.0
pytz==2018.9
scipy==1.2.1
six==1.12.0
subprocess32==3.5.3
